trying to reduce latency of tcp stream

+++ UltraStar Play/Packages/playshared/Runtime/Network/Dto/PositionInSongDto.cs
@@ -2,7 +2,7 @@ public class PositionInSongDto : CompanionAppMessageDto
 {
     public float SongBpm { get; set; }
     public float SongGap { get; set; }
-    public int PositionInSongInMillis { get; set; }
+    public double PositionInSongInMillis { get; set; }
 
     public PositionInSongDto() : base(CompanionAppMessageType.PositionInSong) { }
 }
